backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8

# API & Serialization
marshmallow==3.20.2
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# File Upload & Storage
Pillow==10.1.0
boto3==1.34.0
botocore==1.34.0

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7

# Utilities
requests==2.31.0
python-slugify==8.0.1
bleach==6.1.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Task Queue (optional)
celery==5.3.4
redis==5.0.1

# API Documentation
flasgger==0.9.7.1

# Social Login
python-social-auth==0.3.6
google-auth==2.24.0

# Email Templates
Jinja2==3.1.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Cache
Flask-Caching==2.1.0

# Rate Limiting
Flask-Limiter==3.5.0

# Admin Interface
Flask-Admin==1.6.1

# Search
elasticsearch==8.11.0

# Geolocation
geopy==2.4.0

# Image Processing
opencv-python==4.8.1.78

# Data Export
pandas==2.1.4
openpyxl==3.1.2